Restructure sample to use separate commands
