+++ packages/sample/package.json
@@ -11,7 +11,7 @@
     },
     "scripts": {
         "build": "tsc",
-        "list": "node dist/main.js -l",
-        "led": "node dist/main.js --led"
+        "list": "node dist/main.js list",
+        "led": "node dist/main.js led"
     }
 }
