+++ packages/sample/src/HubStringify.ts
@@ -1,0 +1,17 @@
+import { RevHub } from "@rev-robotics/expansion-hub";
+
+export async function toString(hub: RevHub): Promise<string> {
+    let result = `RevHub: ${hub.moduleAddress}\n`;
+
+    if (hub.isExpansionHub()) {
+        console.log(`Is open? ${hub.isOpen}`);
+    }
+
+    if (hub.isParent()) {
+        for (const child of hub.children) {
+            result += `\tRevHub: ${child.moduleAddress}\n`;
+        }
+    }
+
+    return result;
+}
