+++ packages/sample/src/command/led.ts
@@ -1,0 +1,20 @@
+import {
+    createLedPattern,
+    ExpansionHub,
+    LedPatternStep,
+    openConnectedExpansionHubs,
+} from "@rev-robotics/expansion-hub";
+
+export async function led() {
+    const hubs: ExpansionHub[] = await openConnectedExpansionHubs();
+    const steps = [
+        new LedPatternStep(1, 0, 255, 0), //green
+        new LedPatternStep(1, 255, 0, 0), //red
+        new LedPatternStep(1, 0, 0, 255), //blue
+        new LedPatternStep(1, 255, 0, 255), //magenta
+        new LedPatternStep(1, 255, 255, 0), //yellow
+    ];
+    await hubs[0].sendKeepAlive();
+    const pattern = createLedPattern(steps);
+    await hubs[0].setModuleLedPattern(pattern);
+}
